# lib/api-spec/openapi.orval.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Api",
    "version": "0.2.0",
    "description": "Docuvia — Universal VCS Knowledge Graph Engine API"
  },
  "servers": [{ "url": "/api", "description": "Base API path" }],
...
  "paths": {
...
    "/projects/{id}/pull-requests": {
      "get": {
        "operationId": "listPullRequests",
        "tags": ["github-prs"],
        "summary": "List pull requests for a project",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of pull requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PullRequest" }
                }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/pull-requests/{prNumber}": {
      "get": {
        "operationId": "getPullRequestDetail",
        "tags": ["github-prs"],
        "summary": "Get PR details with knowledge graph impact",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "prNumber", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "PR with impact analysis",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PullRequestDetail" } }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/projects/{id}/pull-requests/{prNumber}/analyze": {
      "post": {
        "operationId": "analyzePullRequest",
        "tags": ["github-prs"],
        "summary": "Manually trigger PR knowledge impact analysis",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "prNumber", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "202": {
            "description": "Analysis triggered",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PrAnalyzeResult" } }
            }
          },
          "404": { "description": "PR not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "L2Node": {
        "type": "object",
        "required": [
          "id",
          "projectId",
          "name",
          "type",
          "l3Count",
          "l1TagIds",
          "aiGenerated",
          "createdAt"
        ],
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["package", "module", "pcd", "sys-uncategorized"] },
          "description": { "type": "string", "nullable": true },
          "l3Count": { "type": "integer" },
          "l1TagIds": { "type": "array", "items": { "type": "integer" } },
          "aiGenerated": { "type": "boolean" },
          "needsReview": { "type": "boolean" },
          "isSystem": { "type": "boolean" },
          "isBootstrapConfirmed": { "type": "boolean" },
          "pathPatterns": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string" }
        }
      },
...
      "L3Node": {
        "type": "object",
        "required": ["id", "l2NodeId", "title", "nodeType", "aiGenerated", "createdAt"],
        "properties": {
          "id": { "type": "integer" },
          "l2NodeId": { "type": "integer" },
          "title": { "type": "string" },
          "content": { "type": "string", "nullable": true },
          "nodeType": { "type": "string", "enum": ["change", "rule", "decision", "context"] },
          "commitHash": { "type": "string", "nullable": true },
          "aiGenerated": { "type": "boolean" },
          "confidence": { "type": "number", "nullable": true },
          "noiseScore": { "type": "number", "nullable": true },
          "createdAt": { "type": "string" }
        }
      },
...
      "PullRequest": {
        "type": "object",
        "required": [
          "id",
          "projectId",
          "githubPrNumber",
          "title",
          "headSha",
          "baseSha",
          "author",
          "state",
          "url",
          "analysisStatus",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "integer" },
          "projectId": { "type": "integer" },
          "githubPrNumber": { "type": "integer" },
          "title": { "type": "string" },
          "body": { "type": "string", "nullable": true },
          "headSha": { "type": "string" },
          "baseSha": { "type": "string" },
          "author": { "type": "string" },
          "state": { "type": "string", "enum": ["open", "closed", "merged"] },
          "url": { "type": "string" },
          "analysisStatus": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed", "failed"]
          },
          "aiSummary": { "type": "string", "nullable": true },
          "mergedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PullRequestDetail": {
        "type": "object",
        "required": ["pr", "commitsCount", "l2Nodes", "l3Nodes"],
        "properties": {
          "pr": { "$ref": "#/components/schemas/PullRequest" },
          "commitsCount": { "type": "integer" },
          "l2Nodes": { "type": "array", "items": { "$ref": "#/components/schemas/L2Node" } },
          "l3Nodes": { "type": "array", "items": { "$ref": "#/components/schemas/L3Node" } },
          "aiSummary": { "type": "string", "nullable": true }
        }
      },
      "PrAnalyzeResult": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": { "type": "string", "enum": ["triggered", "already_completed"] },
          "message": { "type": "string" }
        }
      },
...
    }
  }
}